Moved account menu and news components
Closes #1153
