+++ apps/extension/src/ui/components/account-menu/MainAccountsComponent.tsx
