+++ apps/extension/src/ui/components/account-menu/SwitchAccountCover.tsx
