+++ apps/extension/src/ui/components/news/NewsView.tsx
