+++ apps/extension/src/ui/components/news/news-item.tsx
